Back_End/StudyCon/StudyCon/requirements.txt
Django==5.1.6
daphne==4.0.0
channels==4.0.0
channels-redis==4.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-push-notifications==3.0.0
django-ratelimit==4.1.0
psycopg2-binary
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
bleach==6.1.0
dj-database-url==2.1.0